sample: move duplicated classes to module core:base
