+++ core/base/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *  Copyright (c) 2021 Fraunhofer Institute for Software and Systems Engineering
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,28 +8,27 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Microsoft Corporation - initial API and implementation
+ *       Fraunhofer Institute for Software and Systems Engineering - initial API and implementation
  *
  */
 
 plugins {
     `java-library`
+    `maven-publish`
 }
 
+val jupiterVersion: String by project
 
 dependencies {
     api(project(":spi"))
-    implementation(project(":core:base"))
 
-    testImplementation(project(":extensions:in-memory:transfer-store-memory"))
-
+    testImplementation("org.junit.jupiter:junit-jupiter-api:${jupiterVersion}")
 }
-
 
 publishing {
     publications {
-        create<MavenPublication>("transfer") {
-            artifactId = "transfer"
+        create<MavenPublication>("core-base") {
+            artifactId = "core-base"
             from(components["java"])
         }
     }
