+++ core/contract/build.gradle.kts
@@ -22,6 +22,7 @@ plugins {
 dependencies {
     api(project(":spi"))
     api("org.slf4j:slf4j-api:${slf4jVersion}")
+    api(project(":core:base"))
     implementation(project(":core:policy:policy-engine"))
     testImplementation(project(":extensions:in-memory:negotiation-store-memory"))
 }
